more add helpers
